Refactor(cli): Refactor `Argument` to `Opt` enum
Nim's `parseopt` module defines an "argument" to be everything passed to
the program on the command-line that doesn't start with `-`.

Therefore it was strange to see e.g. `ExerciseArgument` within the
`case` statement `of cmdLongOption, cmdShortOption` rather than
`cmdArgument`.

It is better to rename each old "argument" as an "option", showing that
they start with `-`, and that some of them can take a value.

This commit defines the options as an `enum` with an enum-indexed array
of the keys, instead of a bunch of consts.

We also resolve the issue that `nimpretty` didn't like the old
alignment. See:
https://nim-lang.org/docs/nep1.html#introduction-spacing-and-whitespace-conventions

Note that the name `Opt` avoids a clash with `Option` from the `options`
module.