Fix: Show minimal `git clone` output by default (#60)
Commit f502ab6de239 (#49) caused the program to produce no output during
the `git clone` at the default verbosity setting. It was nice to remove
the excessive `git clone` output, but total removal means that the
program can appear to hang on a poor internet connection.

This commit makes the situation a little better for that case. In the
future, we can try to improve the behaviour further.

Fixes: #52