Fix(cli): Support the syntax of e.g. `-m include` (#54)
This commit adds support for separating a short option key from its
value with a space (which the help message implied was possible). For
example, running either of:
    canonical_data_syncer -m include
    canonical_data_syncer -m i
would previously produce an error:
    Error: '-m' was given without a value

This is because the `parseopt` module in Nim's standard library only
supports separating a short option key from its value with an equals
sign, a colon, or nothing. That is, only these work:
    canonical_data_syncer -m=include
    canonical_data_syncer -m:include
    canonical_data_syncer -m=i
    canonical_data_syncer -m:i
    canonical_data_syncer -mi

But a long option key and its value _can_ be separated by a space (or an
equals sign, or a colon). For example:
    canonical_data_syncer --mode=include
    canonical_data_syncer --mode:include
    canonical_data_syncer --mode include
    canonical_data_syncer --mode=i
    canonical_data_syncer --mode:i
    canonical_data_syncer --mode i

The simplest fix is to use cligen's `parseopt3` module, which is
designed to be API-compatible with `parseopt`. `cligen` is well-known
and mature. Now we can support all of the above syntax, at the cost of:
- One extra dependency at compile-time only.
- Increasing the release binary size by about 2% (the Linux release
  binary is exactly 4 KiB bigger).

Fixes: #45