Refactor(probspecs): Do `grainsWorkaround` at parse-time
We want to add support for specifying the path of a
`problem-specifications` directory to use. However, the previous
workaround for `grains` was implemented by overwriting the
`canonical-data.json` file for `grains`.

It's better if we don't alter a user's non-temporary
`problem-specifications` repo, so this commit changes the `grains`
workaround to happen at the time of parsing instead.