Refactor(cli): Generate help text (#95)
This commit improves how we construct the `--help` string, but does not
change the actual output of `--help`.

Previously, we had to be careful with the spacing so that the help
message for each option is aligned correctly.

This commit also makes it clearer that we set the name of the
application in the help message at run-time, rather than compile-time.