Refactor: Use per-command src subdirectories
This will make the code easier to navigate as we add more commands.

Later on, it will be easier to identify the parts of the current `sync`
code that should be shared between commands - we can then refactor that
code and move it back to the `src` directory.

This commit also tweaks the import style, with imports in this order:
- stdlib imports
- nimble imports
- imports from another directory
- imports from the same directory