Use per-command subdirectories in `src` (#100)
This will make the code easier to navigate as we add more commands.

Later on, it will be easier to identify the parts of the current `sync`
code that should be shared between commands - we can then refactor that
code and move it back to the `src` directory.

This commit also tweaks the import style, with imports in this order:
- Standard library imports - `std/`
- Nimble imports - `pkg/`
- Imports from another directory - typically `".."/`
- Imports from the same directory - `"."/`

Putting square brackets around a single import is probably a bit
idiosyncratic. But maybe it helps readability to make all imports look
the same, and to make diffs slightly less noisy later on.

We can consider adding subdirectories to the `tests` folder later, as we
add more tests.