Implement `uuid` command (#106)
This allows users to run `configlet uuid` to generate a new version 4
UUID.

Compared to the implementation of `configlet uuid` in the configlet for
Exercism v2, this adds a `-n, --num` option for generating multiple
UUIDs.

We currently do not export a proc that returns a new UUID string, but
that might eventually become useful.

This commit also adds tests for malformed commands like:
- `configlet uuid sync` (two commands given)
- `configlet uuid --check` (valid option given to the wrong command)

Note that v0.1.11 is the minimum version of the `uuids` package that
can be used with `--styleCheck:error` on Windows. See upstream fix:
https://github.com/pragmagic/uuids/commit/8cb8720b567c