cli: add `-t, --track-dir` global option (#72)
This commit allows the configlet binary to act on a different directory
than its own. This is useful in general, but it will also help us to
test configlet itself - we can now establish an initial state for
end-to-end tests by specifying:
- a track directory
- a problem-specifications directory
- the offline mode

That allows us to do black-box testing of the release binary, running
something like:
    configlet sync -t=/tmp/python -p=/tmp/my_prob_specs -o -m=include
and then asserting that the changes made are as expected.

Note that we implement `-t, --track-dir` as a global option, rather than
an option for `sync` - it will also be useful for e.g. `lint`.

Some decisions:
- Set the default value of `conf.trackDir` explicitly to the current
  directory in `cli.nim`.
- Determine the `tests.toml` path from `conf.trackDir`, but don't change
  the current directory to `conf.trackDir`.

Closes: #50