build(config): add `--styleCheck:hint` (#166)
This allows us to see styleCheck output if `--verbose` is passed during
`nimble build`. 

The `--verbose` is required because the latest release of Nimble
(0.12.0, in Oct 2020) adds `--hints:off` by default, and that's set
after `--hint[Name]:on` can be passed.

From the Nimble changelog:
> Calls to the Nim compiler now display --hints:off output by default.
> The --verbose flag will print the full Nim command as well as regular
> compiler output.

However, this commit is also useful to turn on styleCheck by default for
a user who runs `nim c foo.nim` in our project. 

`--styleCheck:error` is possible too, but it's currently too aggressive:
it would currently produce an error if one of our Nimble dependencies
has a style problem.

See: #16