lint(validators): refactor `len` conditions
In our validation procs, there are conditions like
```Nim
if data.hasKey(key):
  if data[key].kind == JString:
```
where the condition being `true` represents the "happy path".

To improve readability, let's try to do this consistently for every
condition apart from the innermost one; it's easier to see the "happy
path" when it minimises the number of `else`.

This commit refactors the `len` conditions accordingly, so that we
prefer `len > 0` and use `len == 0` only when there is no `else`.

This commit also changes a `var` to a `let` to better convey that we
don't mutate.