tests: write progress message to stderr, not stdout
We can argue that this message isn't really the "output" of the tests,
but is just a temporary communication to the user that the program
hasn't hung unexpectedly.

`unittest` writes everything to stdout, and so with this commit, the
user who compiles the tests and runs
```
$ ./tests/all_tests > /tmp/configlet_test_results.txt
```
will no longer see the message in that file - they see it in their
terminal instead.