lint(validators): support top-level array (#218)
Previously, we could not lint a JSON file that is an array at the top
level - for example, `links.json`. 

This is a refactoring commit that allows us to do so. We:
- split our `checkArrayOf` proc to `isArrayOf` and `hasArrayOf`.
- split our `checkArrayOfStrings` proc to `isArrayOfStrings` and
  `hasArrayOfStrings`.

There are many ways to refactor and tidy the current linting code
further, but that can wait until a later commit.

I have verified that this commit keeps the output of `configlet lint`
the same for every track.