lint(validators): add `isString`, `isBoolean`, `isInteger`
This commit makes our validation procs consistent with the style of the
existing `hasArray` and `isArray`. This is generally useful, and unlocks
some further refactorings - for example, the checking of `tags` can
use `isString`.

In the future we can make the `hasX` procs use common logic to handle
a missing key.

This commit also removes some duplication between `hasObject` and
`isObject`.