lint(validators): add `isString`, `isBoolean`, `isInteger` (#252)
With this commit, all our JSON validator procs use the `hasX`/`isX`
style, which is consistent with the existing `hasArrayOf`/`isArrayOf`,
and `hasObject`/`isObject`. This is generally useful, and unlocks some
further refactorings - for example, the checking of `tags` can use
`isString`.

In the future we can make the `hasX` procs use common logic to handle
a missing key.

This commit also removes some duplication between `hasObject` and
`isObject`.