lint(track_config): refactor `tags` checking (#253)
The previous commit 489b29042849 (#252) allows us to make this
simplification.

In the future, we could use `hasArrayOfStrings` in `hasValidTags`, which
could allow us to remove the `isValidTag` proc. We can't do this now
because `hasArrayOfStrings` currently cannot check that the string value
is one of a set of allowed strings.