lint: make `path` a `distinct string`
This gives us more type safety, and helps some implementations of the
context handling. For example: it was previously possible to write
`context` and `path` in the wrong order when calling some procs, which
might go unnoticed until it affects an error message.

We can consider doing this across the whole configlet codebase, but
let's keep it within the linting code for now.