lint(validators): enforce `nimcall` for per-item procs
This makes it a compile-time error for one of these `call` procs to
capture a non-global variable. Our existing per-item validator procs
already had the `nimcall` calling convention, but this commit means
we'll notice if we accidentally make one into a `closure`, which has
more overhead.

For more details, see the Nim Manual:
- https://nim-lang.org/docs/manual.html#types-procedural-type