lint(validators): enforce `nimcall` for per-item procs (#269)
This makes it a compile-time error for one of these `call` procs to
capture a non-global variable. Our existing per-item validator procs
already had the `nimcall` calling convention, but this commit guarantees
we can't accidentally make one into a closure. Closures have a
relatively high performance cost in the current version of Nim, and also
tend to be less readable.

For more details, see the Nim Manual:
- https://nim-lang.org/docs/manual.html#types-procedural-type
- https://nim-lang.org/docs/manual.html#procedures-closures