lint, uuid: move and rename UUID validator (#279)
It makes more sense for the UUID validator func to be next to the other
validation code. The func was originally added to `uuid.nim` because
the `lint` command was not yet implemented.

This commit also renames `isValidUuidV4` to `isUuidV4`, which helps to
clarify that it only checks the format of the string - it does not check
that the UUID is unique across Exercism. Previously the name could
mislead a reader - a duplicate UUID can be thought of as "invalid".

The new name is also more consistent with the names of the other
validator procs, which currently omit the word "valid".