lint(validators): support checking arrays for duplicates
Some string arrays have a constraint that their values must be unique, i.e. they must not be repeated.

The `isArrayOfStrings` and `hasArrayOfStrings` functions now take a `uniqueValues` argument with which uniqueness of the array values can be enforced. The default behavior is to allow duplicate values in an array.