lint(validators): support checking arrays for duplicates (#295)
Some string arrays have a constraint that their values must be unique,
i.e. each value must not appear more than once.

The `isArrayOfStrings` and `hasArrayOfStrings` procs now take a
`uniqueValues` argument by which uniqueness of the array values can be
enforced. The default behavior is to allow duplicate values in an array.