ci: use `--styleCheck:hint` not `--styleCheck:error`
`--styleCheck:error` is useful because it makes a identifier naming
inconsistency cause a failing CI check. However, it also checks the
identifier naming in dependencies, such as the Nimble packages that we
import. This has become too inconvenient.

This commit alters our CI to use `nimble test`, which effectively moves
us back to `--styleCheck:hint` (as specified by our `config.nims` file).

In the future, the GitHub Action we use to install Nim should be able to
produce so-called Annotations to make compiler hints more visible,
without causing CI to outright fail.