ci: use `--styleCheck:hint`, not `--styleCheck:error` (#305)
We have been using `--styleCheck:error` so that CI will fail if there is
a style issue in our codebase. However, `--styleCheck:error` also checks
the style in the Nimble packages that we import. This has become too
inconvenient.

This commit alters our CI to use `nimble test`, which effectively moves
us back to `--styleCheck:hint` (as specified by our `config.nims` file).

The development version of Nim has a new command-line option that allows
turning a hint into an error. That is, in the future we could use
`--hintAsError[Name]:on` to get the best of both worlds.