refactor: DRY for parameters
These parameters all have the same default value in `hasArrayOfStrings`,
so we can use that as a single source of truth.

The `hasArrayOfFiles` proc is very specific, and so I'd argue that we
don't need to expose an interface for e.g. allowed values or kebab case
anyway. And e.g. every `isRequired` parameter has the default value of
`true`, so I think we don't need to write it explicitly when calling
`hasArrayOfStrings` either.

We can add `uniqueValues = true` if a future update of the spec dictates
as such.