generate: add initial `generate` command (#308)
Each Concept Exercise directory must have a `.docs/introduction.md`
file, which introduces the concept(s) that the exercise teaches.
However, a track may want that file to contain the contents of a
Concept's `introduction.md` file.

To address the duplication, this commit adds the `generate` command to
configlet. A track can now add a template file named
`introduction.md.tpl` in a Concept Exercise's `.docs` directory, and run
`configlet generate` to write the `introduction.md` file in the same
directory. The generated file is just the contents of the template file,
but any instance of `%{concept:some-slug}` is replaced by the contents
of the `introduction.md` file for the Concept `some-slug`.

For example, here is a valid `introduction.md.tpl` file for a track that
has a `recursion` Concept:

```
# Recursion

%{concept:recursion}
```

For now, the generated file does not contain the first level-one heading
of the Concept's `introduction.md` file. This gives a track more
flexibility - for example, they can use a different level-one heading.

`configlet generate` does not yet support specifying a single exercise.
In the future it should probably support the same `-e, --exercise`
option as the `sync` command.

For more information, see the docs:
- https://github.com/exercism/docs/blob/fed4c0f59ece/building/tracks/concept-exercises.md#file-docsintroductionmdtpl

Co-authored-by: ee7 <45465154+ee7@users.noreply.github.com>