sync(exercises, tracks): refactor (#332)
This commit is part of a series that tries to improve readability and
maintainability of the sync code. It also improves performance slightly;
`configlet sync --check` is about 5% faster. Future commits will
continue refactoring, and provide much larger performance improvements.

This is a squashed commit. Summary:
- deduplicate `testsFile` logic
- rename `trackExercise` to `practiceExercise`
- inline second `findPracticeExercises` proc for readability
- sort inside `findPracticeExercises`
- use `PracticeExerciseSlug`, not `practiceExercisePath`
- add doc comments
- shorten some long lines
- remove unused `hasCanonicalData` func
- allocate less

See #332 for the (granular) pre-squash commits.

I've verified that this commit keeps the output of
`configlet sync --check` the same on every track.

See also:
- #64, which tracks the refactoring of the sync code.