sync: refactor prob-specs handling, and speed-up (#333)
This commit is part of a series that tries to improve readability and
maintainability of the sync code. This time, we focus on the
`probspecs.nim` code, and also improve performance: an offline
`configlet sync --check` is about 2.8 - 4.⁠5x faster, depending on the
track.

Measuring the execution time of
  configlet sync --check --offline --prob-specs-dir=path/to/problem-specifications

in different track repos, before and after this commit:

|                 |            ada |            nim |           csharp |
| --------------- | -------------- | -------------- | ---------------- |
| previous commit |  92.2 ± 0.9 ms | 103.5 ± 1.0 ms |   111.0 ± 1.1 ms |
| this commit     |  20.6 ± 0.7 ms |  31.8 ± 0.9 ms |    39.3 ± 1.0 ms |

These tracks have respectively 0, 67, and 112 exercises with canonical
data. Each timing value is the average wall-clock time per run
(mean ± standard deviation) for 100 runs (top row) and 500 runs
(bottom row) respectively.

The previous slowness didn't badly affect the end user, but it was
particularly noticeable when running `configlet sync --check` on every
track for regression testing.

With this commit, most of the runtime of an offline
`configlet sync --check` is now spent parsing TOML and JSON, as we'd
expect. See the profiling graph in #333 for more details.

An upcoming commit will optimize further: the first implementation of
`sync` just parsed every `canonical-data.json` file in
`problem-specifications` regardless of the track. But e.g. if a track
has only one exercise with canonical data, we should only parse the
`canonical-data.json` file for that one exercise.

After that optimization, the remaining slowness is because we
currently parse the JSON/TOML into a dynamic structure
(`JsonNode`/`TomlTable`), rather than directly populating a
strongly typed object. We'll keep this approach, for now - it's simplest
to stay with the `std/json` and `pkg/parsetoml` modules for the early
implementation.

This is a squashed commit. Summary:
- allocate less: avoid intermediate sequences, and make
  `exercisesWithCanonicalData` an `iterator`.
- prefer `distinct string` to an `object` with one `string` field
- prefer `distinct JsonNode` to an `object` with one `JsonNode` field`
- return a `Table` from `findProbSpecsExercies` - previously we
  immediately transformed the returned object into a `Table`.
- combine procedures where it seems to reduce cognitive overhead
- improve and rename `execCmdException`
- add some doc comments
- prefer `func` where possible
- add more `requiresInit`
- tweak style

See #333 for the (granular) pre-squash commits.

I've verified that this commit keeps the output of
`configlet sync --check` the same on every track.

See also:
- #64, which tracks the refactoring of the sync code.