sync: clarify call vs field access
Again, this isn't necessarily my preferred style. But we've done it
elsewhere, and it does add syntax highlighting in most editors and on
GitHub.

My main gripe was with `.status` - it wouldn't be immediately
obvious to a new reader (or possibly me in the future) whether this is
a call or a field access.