sync(sync): remove `doAssert`s
We haven't used this kind of checking elsewhere in our codebase, and we
only call these procs like:

```
  if testCase.reimplements.isNone():
    chooseRegularSyncDecision(testCase)
  else:
    chooseReimplementsSyncDecision(testCase)
```

so I think we can do without these checks. And even with this commit,
we'd still raise an `UnpackDefect` if we incorrectly alter the code so
that we try to `get` from a `none`.

Furthermore, if there's a real danger of this check being `false` then
we should probably use an `if` or `try` so that the user gets a better
error message.