sync: don't parse unneeded `canonical-data.json` files
Running `configlet sync` would previously parse every
`canonical-data.json` file in the problem-specifications repo. There are
currently 119 of these files, and the execution time of `configlet sync`
is usually dominated by JSON parsing.

With this commit, we only parse a `canonical-data.json` file if the
track implements the corresponding exercise - this is a large speedup
for tracks that implement only a small fraction of the exercises in
`problem-specifications`.