tests(binary): speed up local testing
Before running our tests of the configlet executable, our setup
automatically builds configlet (so that we don't forget).

We should test an executable that is as close as possible to that we
distribute, which means we should build with the same compilation flags.
However, a debug build should usually produce the same test status -
with this commit, we perform the longer build only during CI so that
running the tests locally is faster.