tests(binary): speed up local testing (#344)
Before running our tests of the configlet executable, our setup
automatically builds configlet (so that we don't forget).

We should test an executable that is as close as possible to that we
distribute, which means we should build with the same compilation flags.
However, a debug build should usually produce the same test status -
with this commit, we perform the longer build only during CI so that
running the tests locally is faster.

On a slow machine, with an already-existing `nimcache`, recompiling
after a trivial code change takes about 2 seconds for the debug build,
but more like 10 seconds for the release build. Nearly all of this
difference is due to the `--passC:-flto` flag.