sync(exercises): speed up writing of `tests.toml` (#359)
This commit improves performance of `configlet sync --update` by fixing
an oversight. The improvement is generally not noticeable during typical
use.

Previously, when writing a `tests.toml` file, we parsed the existing
`tests.toml` file once per test case that we want to include.

On a slow machine, this commit makes an offline non-interactive sync via

    configlet sync --offline -p my-local-ps-dir --update --mode include

finish in about 20 ms, rather than about 35 ms previously. This data
point is for a track with 67 exercises that have canonical data, where
10 exercises are updated.