tests(binary): support skipping build (#391)
The tests of the configlet executable automatically run `nimble build`
first, so that we don't forget to. This is a fine default, but makes it
significantly slower to re-run those tests when making changes to them
alone.

This commit allows a configlet developer to run

    nim r -d:skipBuild tests/test_binary

to skip the `nimble build` when it is unnecessary.