lint(track_config): fix handling of missing `status` (#414)
A Concept Exercise or Practice Exercise in the track-level
`config.json` file may have a `status` property. If that property is
missing, the implied value is `active`, and so the exercise appears on
the website.

With this commit, `configlet lint` no longer skips some checks of a
Concept Exercise when its `status` property is missing.