lint(track_config): improve jsony error message
Before this commit, an error during jsony's deserialization would look
something like:

  jsony.nim(46)            parseHook
  Error: unhandled exception: Expected " but got 1 instead. At offset: 73830 [JsonError]

With this commit, it instead looks like:

  Error: JSON parsing error during the second linting pass:
  config.json(3156, 10): Expected " but got 1 instead.

That is, we:
- Handle the exception
- Better explain what happened
- Mention the JSON file being parsed, rather than `jsony.nim`
- Provide the line and column number, rather than the offset, where the
  parsing error occurred

However, we hope that a user shouldn't encounter these errors, which
should only occur when a value's type in the `config.json` file doesn't
match that in the `TrackConfig` object. The current approach tries to
validate the types in the first pass, and only perform the second pass
when the types are correct.

Currently `showError` also prints the help message. We'll add a better
error proc in the future (maybe in `helpers.nim`, not `cli.nim`).