lint(track_config): tweak `hasObject` use
It was a little confusing that passing `isRequired = false` to
`hasObject` did nothing, because we checked `hasKey` twice.

But the problem is that the existing `hasX` design is a bit of a mess,
or at least the proc names. I don't love that a proc named `hasObject`
returns true when `isRequired = false` is passed.

Let's just do it this way for now - I found it little clearer.
Eventually, all this `hasX` code might be refactored.