sync: make some funcs proc
This change is required for upgrading to Nim 1.6.0.

In Nim 1.4.x and Nim 1.6.0 rc2, compiling the below code with
`--experimental:strictFuncs` produces an error saying that `[]` has
side effects.

    func getFooVal(j: JsonNode): JsonNode =
      j["foo"]

However, in Nim 1.4.x, there was no error produced when similarly
compiling the below

    type
      MyNode = distinct JsonNode

    proc `[]`(j: MyNode, name: string): MyNode {.borrow.}

    func getFooVal(j: MyNode): MyNode =
      j["foo"]

Nim 1.6.0 _does_ produce an error in the latter case, so that the side
effect error is produced for using `[]` with both a `JsonNode` and a
`distinct JsonNode`.

This means we need to change some `func` to `proc` in our codebase,
because we use

    ProbSpecsTestCase* = distinct JsonNode

That is, compiling the below with `--experimental:strictFuncs` did not
produce an error with Nim 1.4.8, but does produce an error with
Nim 1.6.0.

    func uuid*(testCase: ProbSpecsTestCase): string =
      testCase["uuid"].getStr()