github: set configlet release as draft initially (#444)
Before this commit, a configlet release was fully published as soon as
the first build job completed. This caused several problems - the most
severe being that a transient error in the Linux build would cause
`configlet lint` to fail in CI Exercism-wide, because the
`fetch-configlet` script fails to download the missing Linux asset from
the incomplete release.

With this commit, we set the release as a draft initially. Now, the
`fetch-configlet` script does not attempt to download assets from that
release until we manually un-draft it. This is because the
GitHub Releases API endpoint 

    GET https://api.github.com/repos/{owner}/{repo}/releases/latest

returns "the most recent non-prerelease, non-draft release, sorted by
the `created_at` attribute".

It would be possible to automatically un-draft the release by checking
whether the total number of assets equals the expected final number of
assets (currently 3), via something like:

    gh release view --json assets | jq -r '.assets | length'  

But let's just un-draft manually - it's more robust, and not
inconvenient.

Closes: https://github.com/exercism/configlet/issues/223
Closes: https://github.com/exercism/github-actions/issues/19